requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
amqp==5.2.0
    # via kombu
annotated-types==0.6.0
    # via pydantic
asgiref==3.7.2
    # via django
billiard==4.2.0
    # via celery
celery[redis]==5.3.6
    # via
    #   -r requirements.in
    #   django-celery-beat
certifi==2023.11.17
    # via
    #   requests
    #   sentry-sdk
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.3.0
    # via celery
cron-descriptor==1.4.0
    # via django-celery-beat
dj-database-url==2.1.0
    # via -r requirements.in
django==4.2.9
    # via
    #   -r requirements.in
    #   dj-database-url
    #   django-celery-beat
    #   django-formtools
    #   django-ninja
    #   django-otp
    #   django-phonenumber-field
    #   django-sesame
    #   django-timezone-field
    #   django-two-factor-auth
    #   sentry-sdk
django-celery-beat==2.5.0
    # via -r requirements.in
django-formtools==2.5.1
    # via django-two-factor-auth
django-ninja==1.1.0
    # via -r requirements.in
django-otp==1.3.0
    # via django-two-factor-auth
django-phonenumber-field==7.3.0
    # via django-two-factor-auth
django-sesame==3.2.2
    # via -r requirements.in
django-timezone-field==6.1.0
    # via django-celery-beat
django-two-factor-auth[phonenumberslite]==1.15.5
    # via -r requirements.in
django-webpack-loader==3.0.0
    # via -r requirements.in
gunicorn==21.2.0
    # via -r requirements.in
idna==3.6
    # via requests
jinja2==3.1.3
    # via -r requirements.in
kombu==5.3.5
    # via celery
markupsafe==2.1.3
    # via jinja2
packaging==23.2
    # via gunicorn
phonenumberslite==8.13.27
    # via django-two-factor-auth
prompt-toolkit==3.0.43
    # via click-repl
psycopg2==2.9.9
    # via -r requirements.in
pydantic==2.5.3
    # via django-ninja
pydantic-core==2.14.6
    # via pydantic
pypng==0.20220715.0
    # via qrcode
python-crontab==3.0.0
    # via django-celery-beat
python-dateutil==2.8.2
    # via
    #   celery
    #   python-crontab
qrcode==7.4.2
    # via django-two-factor-auth
redis==5.0.1
    # via
    #   -r requirements.in
    #   celery
requests==2.31.0
    # via stripe
sentry-sdk[django]==1.39.2
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
sqlparse==0.4.4
    # via django
stripe==7.12.0
    # via -r requirements.in
typing-extensions==4.9.0
    # via
    #   dj-database-url
    #   pydantic
    #   pydantic-core
    #   qrcode
    #   stripe
tzdata==2023.4
    # via
    #   celery
    #   django-celery-beat
urllib3==2.1.0
    # via
    #   requests
    #   sentry-sdk
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.13
    # via prompt-toolkit
whitenoise==6.6.0
    # via -r requirements.in
